# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TontiTOGO API",
        "description": "API RESTful de la plateforme TontiTOGO — Gestion numérique des tontines au Togo avec suivi des cotisations et notifications SMS.",
        "contact": {
            "name": "Support TontiTOGO",
            "email": "support@tontitogo.tg"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
...
        "/api/agent/contributions": {
            "get": {
                "tags": [
                    "Agent - Cotisations"
                ],
                "summary": "Historique des cotisations de l'agent (infinite scroll, 20 par page)",
                "operationId": "b946357991b6f00b559c42464fd504cb",
                "parameters": [
                    {
                        "name": "settlement_status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "settled"
                            ]
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée des cotisations de l'agent"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Agent - Cotisations"
                ],
                "summary": "Enregistrer une cotisation terrain",
                "operationId": "e479ce90aa7b75c83c3824e07a4a44d7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tontine_participant_id",
                                    "amount"
                                ],
                                "properties": {
                                    "tontine_participant_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 5000
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 6.1375
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1.2123
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cotisation enregistrée + SMS envoyé"
                    },
                    "422": {
                        "description": "Erreur métier ou validation"
                    },
                    "403": {
                        "description": "Tontine hors organisation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/agent/contributions/{id}": {
            "get": {
                "tags": [
                    "Agent - Cotisations"
                ],
                "summary": "Détail d'une cotisation de l'agent",
                "operationId": "ba40286d0420c4ffea8ff75b173532a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail de la cotisation"
                    },
                    "404": {
                        "description": "Non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez votre token Sanctum : Bearer {token}"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}